move http tests to http package. use lowercase for multi_value_mode enum as the test data is now copied anyway fix existing gauss scorer test
