+++ elastic4s-http/src/test/resources/gauss_scorer.json
@@ -6,7 +6,7 @@
         "boost_mode": "multiply",
         "functions": [
             {
-                "gaussian": {
+                "gauss": {
                     "myfield": {
                         "origin": "now",
                         "scale": "28d",
