+++ elastic4s-http/src/test/resources/score_fieldvaluefactor.json
@@ -1,0 +1,8 @@
+{
+  "field_value_factor": {
+    "field": "likes",
+    "factor": 1.2,
+    "modifier": "sqrt",
+    "missing": 1.0
+  }
+}
