+++ elastic4s-http/src/test/resources/score_random.json
@@ -1,0 +1,5 @@
+{
+    "random_score": {
+        "seed": 12345
+    }
+}
