+++ elastic4s-http/src/test/resources/score_weight.json
@@ -1,0 +1,3 @@
+{
+    "weight": 1.5
+}
